Move initialise logic to root.
